Break up structure a little (#10)
